resolved all conflicts

+++ testsuites/CBLTester/CBL_Functional_tests/conftest.py
@@ -17,6 +17,7 @@ from keywords.exceptions import ProvisioningError
 from keywords.tklogging import Logging
 from keywords.constants import RESULTS_DIR
 
+from CBLClient.FileLogging import FileLogging
 from CBLClient.Replication import Replication
 from CBLClient.BasicAuthenticator import BasicAuthenticator
 from CBLClient.Database import Database
@@ -52,7 +53,8 @@ def pytest_addoption(parser):
 
     parser.addoption("--sync-gateway-version",
                      action="store",
-                     help="sync-gateway-version: Sync Gateway version to install (ex. 1.3.1-16 or 590c1c31c7e83503eff304d8c0789bdd268d6291)")
+                     help="sync-gateway-version: Sync Gateway version to install "
+                          "(ex. 1.3.1-16 or 590c1c31c7e83503eff304d8c0789bdd268d6291)")
 
     parser.addoption("--liteserv-platform",
                      action="store",
@@ -115,6 +117,23 @@ def pytest_addoption(parser):
     parser.addoption("--debug-mode", action="store_true",
                      help="Enable debug mode for the app ", default=False)
 
+    parser.addoption("--use-views",
+                     action="store_true",
+                     help="If set, uses views instead of GSI - SG 2.1 and above only")
+
+    parser.addoption("--number-replicas",
+                     action="store",
+                     help="Number of replicas for the indexer node - SG 2.1 and above only",
+                     default=0)
+
+    parser.addoption("--enable-file-logging",
+                     action="store_true",
+                     help="If set, CBL file logging would enable. Supported only cbl2.5 onwards")
+
+    parser.addoption("--delta-sync",
+                     action="store_true",
+                     help="delta-sync: Enable delta-sync for sync gateway")
+
 
 # This will get called once before the first test that
 # runs with this as input parameters in this file
@@ -144,6 +163,10 @@ def params_from_base_suite_setup(request):
     ci = request.config.getoption("--ci")
     debug_mode = request.config.getoption("--debug-mode")
     no_conflicts_enabled = request.config.getoption("--no-conflicts")
+    use_views = request.config.getoption("--use-views")
+    number_replicas = request.config.getoption("--number-replicas")
+    delta_sync_enabled = request.config.getoption("--delta-sync")
+    enable_file_logging = request.config.getoption("--enable-file-logging")
 
     test_name = request.node.name
 
@@ -182,7 +205,6 @@ def params_from_base_suite_setup(request):
 
     cluster_utils = ClusterKeywords(cluster_config)
     cluster_topology = cluster_utils.get_cluster_topology(cluster_config)
-    cluster_topology = cluster_utils.get_cluster_topology(cluster_config)
 
     sg_url = cluster_topology["sync_gateways"][0]["public"]
     sg_ip = host_for_url(sg_url)
@@ -196,9 +218,6 @@ def params_from_base_suite_setup(request):
         persist_cluster_config_environment_prop(cluster_config, 'sync_gateway_ssl', True)
         target_url = "wss://{}:4984/{}".format(sg_ip, sg_db)
         target_admin_url = "wss://{}:4985/{}".format(sg_ip, sg_db)
-
-    cbs_url = cluster_topology['couchbase_servers'][0]
-    cbs_ip = host_for_url(cbs_url)
 
     if sg_lb:
         persist_cluster_config_environment_prop(cluster_config, 'sg_lb_enabled', True)
@@ -239,6 +258,24 @@ def params_from_base_suite_setup(request):
         log_info("Running with allow conflicts")
         persist_cluster_config_environment_prop(cluster_config, 'no_conflicts_enabled', False)
 
+    if use_views:
+        log_info("Running SG tests using views")
+        # Enable sg views in cluster configs
+        persist_cluster_config_environment_prop(cluster_config, 'sg_use_views', True)
+    else:
+        log_info("Running tests with cbs <-> sg ssl disabled")
+        # Disable sg views in cluster configs
+        persist_cluster_config_environment_prop(cluster_config, 'sg_use_views', False)
+
+    if delta_sync_enabled:
+        log_info("Running with delta sync")
+        persist_cluster_config_environment_prop(cluster_config, 'delta_sync_enabled', True)
+    else:
+        log_info("Running without delta sync")
+        persist_cluster_config_environment_prop(cluster_config, 'delta_sync_enabled', False)
+
+    # Write the number of replicas to cluster config
+    persist_cluster_config_environment_prop(cluster_config, 'number_replicas', number_replicas)
     cluster_utils = ClusterKeywords(cluster_config)
     cluster_topology = cluster_utils.get_cluster_topology(cluster_config)
     cbs_url = cluster_topology['couchbase_servers'][0]
@@ -262,7 +299,7 @@ def params_from_base_suite_setup(request):
             logging_helper.fetch_and_analyze_logs(cluster_config=cluster_config, test_name=request.node.name)
             raise
 
-    # Hit this intalled running services to verify the correct versions are installed
+    # Hit this installed running services to verify the correct versions are installed
     cluster_utils.verify_cluster_versions(
         cluster_config,
         expected_server_version=server_version,
@@ -278,12 +315,22 @@ def params_from_base_suite_setup(request):
         log_info("Starting TestServer...")
         test_name_cp = test_name.replace("/", "-")
         if device_enabled:
-            testserver.start_device("{}/logs/{}-{}-{}.txt".format(RESULTS_DIR, type(testserver).__name__, test_name_cp, datetime.datetime.now()))
+            testserver.start_device("{}/logs/{}-{}-{}.txt".format(RESULTS_DIR, type(testserver).__name__,
+                                                                  test_name_cp,
+                                                                  datetime.datetime.now()))
         else:
-            testserver.start("{}/logs/{}-{}-{}.txt".format(RESULTS_DIR, type(testserver).__name__, test_name_cp, datetime.datetime.now()))
+            testserver.start("{}/logs/{}-{}-{}.txt".format(RESULTS_DIR, type(testserver).__name__,
+                                                           test_name_cp,
+                                                           datetime.datetime.now()))
 
     suite_source_db = None
+    suite_db = None
     if create_db_per_suite:
+        if enable_file_logging and liteserv_version >= "2.5.0":
+            cbllog = FileLogging(base_url)
+            cbllog.configure(log_level="verbose", max_rotate_count=2,
+                             max_size=1000 * 512, plain_text=True)
+            log_info("Log files available at - {}".format(cbllog.get_directory()))
         # Create CBL database
         suite_cbl_db = create_db_per_suite
         suite_db = Database(base_url)
@@ -306,7 +353,7 @@ def params_from_base_suite_setup(request):
 
         log_info("Loading sample bucket {}".format(enable_sample_bucket))
         server.load_sample_bucket(enable_sample_bucket)
-        server._create_internal_rbac_bucket_user(enable_sample_bucket)
+        server._create_internal_rbac_bucket_user(enable_sample_bucket, cluster_config=cluster_config)
 
         # Restart SG after the bucket deletion
         sync_gateways = cluster_topology["sync_gateways"]
@@ -331,7 +378,9 @@ def params_from_base_suite_setup(request):
         # Create primary index
         password = "password"
         log_info("Connecting to {}/{} with password {}".format(cbs_ip, enable_sample_bucket, password))
-        sdk_client = Bucket('couchbase://{}/{}'.format(cbs_ip, enable_sample_bucket), password=password, timeout=SDK_TIMEOUT)
+        sdk_client = Bucket('couchbase://{}/{}'.format(cbs_ip, enable_sample_bucket),
+                            password=password,
+                            timeout=SDK_TIMEOUT)
         log_info("Creating primary index for {}".format(enable_sample_bucket))
         n1ql_query = 'create primary index on {}'.format(enable_sample_bucket)
         query = N1QLQuery(n1ql_query)
@@ -340,7 +389,7 @@ def params_from_base_suite_setup(request):
         # Start continuous replication
         repl_obj = Replication(base_url)
         auth_obj = BasicAuthenticator(base_url)
-        authenticator = auth_obj.create("traveL-sample", "password")
+        authenticator = auth_obj.create("travel-sample", "password")
         repl_config = repl_obj.configure(source_db=suite_source_db,
                                          target_url=target_admin_url,
                                          replication_type="PUSH_AND_PULL",
@@ -348,7 +397,8 @@ def params_from_base_suite_setup(request):
                                          replicator_authenticator=authenticator)
         repl = repl_obj.create(repl_config)
         repl_obj.start(repl)
-        repl_obj.wait_until_replicator_idle(repl)
+        # max_times is 3000 to give more time to replicate travel sample as it is huge
+        repl_obj.wait_until_replicator_idle(repl, max_times=3000)
         log_info("Stopping replication")
         repl_obj.stop(repl)
 
@@ -375,7 +425,9 @@ def params_from_base_suite_setup(request):
         "sg_config": sg_config,
         "testserver": testserver,
         "device_enabled": device_enabled,
-        "flush_memory_per_test": flush_memory_per_test
+        "flush_memory_per_test": flush_memory_per_test,
+        "delta_sync_enabled": delta_sync_enabled,
+        "enable_file_logging": enable_file_logging
     }
 
     if create_db_per_suite:
@@ -385,18 +437,19 @@ def params_from_base_suite_setup(request):
         suite_db.deleteDB(suite_source_db)
         time.sleep(1)
 
-    # Flush all the memory contents on the server app
-    log_info("Flushing server memory")
-    utils_obj = Utils(base_url)
-    utils_obj.flushMemory()
-
+    if create_db_per_suite:
+        # Flush all the memory contents on the server app
+        log_info("Flushing server memory")
+        utils_obj = Utils(base_url)
+        utils_obj.flushMemory()
+        log_info("Stopping the test server per suite")
+        testserver.stop()
     # Delete png files under resources/data
     clear_resources_pngs()
 
 
 @pytest.fixture(scope="function")
 def params_from_base_test_setup(request, params_from_base_suite_setup):
-    base_url = params_from_base_suite_setup["base_url"]
     cluster_config = params_from_base_suite_setup["cluster_config"]
     xattrs_enabled = params_from_base_suite_setup["xattrs_enabled"]
     liteserv_host = params_from_base_suite_setup["liteserv_host"]
@@ -418,15 +471,19 @@ def params_from_base_test_setup(request, params_from_base_suite_setup):
     liteserv_platform = params_from_base_suite_setup["liteserv_platform"]
     testserver = params_from_base_suite_setup["testserver"]
     device_enabled = params_from_base_suite_setup["device_enabled"]
-    flush_memory_per_test = params_from_base_suite_setup["flush_memory_per_test"]
     enable_sample_bucket = params_from_base_suite_setup["enable_sample_bucket"]
+    liteserv_version = params_from_base_suite_setup["liteserv_version"]
+    delta_sync_enabled = params_from_base_suite_setup["delta_sync_enabled"]
+    enable_file_logging = params_from_base_suite_setup["enable_file_logging"]
+
     source_db = None
-    cbl_db = None
+    test_name_cp = test_name.replace("/", "-")
+    log_filename = "{}/logs/{}-{}-{}.txt".format(RESULTS_DIR, type(testserver).__name__,
+                                                 test_name_cp,
+                                                 datetime.datetime.now())
 
     if create_db_per_test:
         log_info("Starting TestServer...")
-        test_name_cp = test_name.replace("/", "-")
-        log_filename = "{}/logs/{}-{}-{}.txt".format(RESULTS_DIR, type(testserver).__name__, test_name_cp, datetime.datetime.now())
         if device_enabled:
             testserver.start_device(log_filename)
         else:
@@ -445,7 +502,13 @@ def params_from_base_test_setup(request, params_from_base_suite_setup):
     db_config = None
 
     db = None
+    cbl_db = None
     if create_db_per_test:
+        if enable_file_logging and liteserv_version >= "2.5.0":
+            cbllog = FileLogging(base_url)
+            cbllog.configure(log_level="verbose", max_rotate_count=2,
+                             max_size=1000 * 512, plain_text=True)
+            log_info("Log files available at - {}".format(cbllog.get_directory()))
         cbl_db = create_db_per_test + str(time.time())
         # Create CBL database
         db = Database(base_url)
@@ -485,7 +548,9 @@ def params_from_base_test_setup(request, params_from_base_suite_setup):
         "testserver": testserver,
         "db_config": db_config,
         "enable_sample_bucket": enable_sample_bucket,
-        "log_filename": log_filename
+        "log_filename": log_filename,
+        "liteserv_version": liteserv_version,
+        "delta_sync_enabled": delta_sync_enabled
     }
 
     log_info("Tearing down test")
@@ -493,15 +558,20 @@ def params_from_base_test_setup(request, params_from_base_suite_setup):
         # Delete CBL database
         log_info("Deleting the database {} at test teardown".format(create_db_per_test))
         time.sleep(1)
-        db.deleteDB(source_db)
-
-    if flush_memory_per_test:
+        path = db.getPath(source_db).rstrip("/\\")
+        if '\\' in path:
+            path = '\\'.join(path.split('\\')[:-1])
+        else:
+            path = '/'.join(path.split('/')[:-1])
+        if db.exists(cbl_db, path):
+            db.deleteDB(source_db)
         log_info("Flushing server memory")
         utils_obj = Utils(base_url)
         utils_obj.flushMemory()
 
-    log_info("Stopping the test server")
-    testserver.stop()
+    if create_db_per_test:
+        log_info("Stopping the test server per test")
+        testserver.stop()
 
 
 @pytest.fixture(scope="class")
@@ -511,7 +581,6 @@ def class_init(request, params_from_base_suite_setup):
 
     db_obj = Database(base_url)
     doc_obj = Document(base_url)
-    dict_obj = Dictionary(base_url)
     datatype = DataTypeInitiator(base_url)
     repl_obj = Replication(base_url)
     array_obj = Array(base_url)
